jpeg baseline

+++ src/image.rs
@@ -2,17 +2,17 @@
 // by Desmond Germans, 2019
 
 pub struct Image {
-    pub width: u32,
-    pub height: u32,
+    pub width: usize,
+    pub height: usize,
     pub data: Vec<u32>,
 }
 
 impl Image {
-    pub fn new(width: u32,height: u32) -> Image {
+    pub fn new(width: usize,height: usize) -> Image {
         Image {
             width: width,
             height: height,
-            data: vec![0; (width * height) as usize],
+            data: vec![0; width * height],
         }
     }
 }
